split ptable module by backend into matplotlib and much smaller plotly
