+++ pymatviz/ptable/_process_data.py
