+++ pymatviz/utils.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 import warnings
+from collections.abc import Sequence
 from contextlib import contextmanager
 from functools import partial, wraps
 from os.path import dirname
@@ -17,7 +18,7 @@ from matplotlib.offsetbox import AnchoredText
 
 
 if TYPE_CHECKING:
-    from collections.abc import Generator, Sequence
+    from collections.abc import Generator
     from typing import ParamSpec, TypeVar
 
     from numpy.typing import ArrayLike
@@ -49,6 +50,7 @@ CrystalSystem = Literal[
 
 elements_csv = f"{ROOT}/pymatviz/elements.csv"
 df_ptable = pd.read_csv(elements_csv, comment="#").set_index("symbol")
+ElemValues = Union[dict[Union[str, int], float], pd.Series, Sequence[str]]
 
 # fallback value (in nanometers) for covalent radius of an element
 # see https://wikipedia.org/wiki/Atomic_radii_of_the_elements
